End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../models/message';
 import { Pagination } from '../models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) {}
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) {}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -43,10 +44,15 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: () => {
-        this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+    // show a confirm window if the user wants to delete a message
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone!').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe({
+          next: () => {
+            this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+          }
+        });
       }
-    })
+    });
   }
 }
